added benjie's patch for Mac OS X compile.
git-svn-id: svn://okws.org/trunk/okws2@631 7d287422-6eea-da11-b677-00123f2a783a

+++ libaok/oklog.h
@@ -99,7 +99,7 @@ public:
   inline logbuf_t &status_long (u_int st)
   { return put (st).spc ().copy (http_status[st]); }
   inline logbuf_t &uid (u_int64_t i) { return put (i); }
-  inline logbuf_t &inflated_len (size_t l) { return put (l); }
+  inline logbuf_t &inflated_len (size_t l) { return put ((unsigned int)l); }
 
   inline logbuf_t &bcpy (const char *c, u_int len);
   inline logbuf_t &qcpy (const char *c, u_int len);
